Refactor: centralize validateAndNormalize for frontend
