+++ src/pages/PlaySee.tsx
@@ -4,7 +4,7 @@ import { WordCard } from "../components/WordCard";
 import { useGameDeck } from "../hooks/useGameDeck";
 import { useProgress } from "../state/useProgress";
 import { loadAllPacksSafe } from "../data/loadAllPacksSafe";
-import type { WordEntry } from "../data/validateAndNormalize";
+import type { WordEntry } from "@/utils/validateAndNormalize";
 import { ProgressBar } from "../components/ProgressBar";
 import { useToast } from "../state/useToast";
