+++ src/types.ts
@@ -1,2 +1,2 @@
-export type { WordEntry } from './data/validateAndNormalize';
+export type { WordEntry } from '@/utils/validateAndNormalize';
 export type { Pack } from './data/loadAllPacksSafe';
